Move from configuration to install
